+++ DataStructuresProject3/Random.h
@@ -7,6 +7,6 @@ private:
 	static std::mt19937 random;
 public:
 	static int Int(int min, int max);
-	static int Float(float min, float max);
+	static float Float(float min, float max);
 };
